back-end/requirements.txt
pandas
numpy
matplotlib
scikit-learn
tensorflow
opencv-python
seaborn
Flask
Keras
Pillow